stop automatic redirect to `ward/lvisits`
the redirect happens based on the token which has been set on the cookie.
The token is generated based on the code provided. If the code is the
is the same for a ward login as it is for the admin login then the
cookie will be set for both. This is because the ward login does not
require a password. The same applies for when the trust admin
code is the same as the ward code. So in both cases a token for a ward
is set which permits a redirect to the ward visits page.  Therefore a guard clause is set to check if a
cookie is set for a trsut admin or an admin before setting it for the
ward login.